fix: return 404 instead of 500 if an MPA action has an unrecognized id
If an action id is not found in the module map, `decodeAction` throws:
```
Error: Could not find the module "{id}" in the React Server Manifest. This is probably a bug in the React Server Components bundler.
```
Previously, this error would bubble up from the action handler and result in a 500.
This is inconsistent with fetch actions, where we print a special warning and return a 404.

To distinguish "action not found" from other error causes (e.g. a malformed request),
I wrap the server action manifest in a proxy that throws a special error if the action is missing.
We can then detect this error and handle it separately.